Form a list Containing SCF Data of Each Optimization Cycles

This function reads a Gaussian .log file and outputs a list of tibbles, each of which is the SCF Data of a optimization step.

Arguments

directory

A string vector describing the directory of the Gaussian log file.

Value

A list of lists. First lists is a list of tibble, each element in the list refers to a tibble recording the SCF Data of a optimization step. The second list only have one element that is a numeric vector refering to the SCF convergence requirement read from log file.

Examples

library(readr)
library(stringr)
library(tidyselect)
library(utils)
library(dplyr)
library(tibble)

OptiSCFMonitorAsList(SCFMonitorExample())
#> [[1]]
#> [[1]][[1]]
#> # A tibble: 49 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.50D-02 9.10D+00 NA        4.62D-01
#>  2     2 2.31D-02 9.22D+00 -5.53D-01 1.79D-01
#>  3     3 1.22D-02 3.79D+00 -3.03D-01 1.19D-01
#>  4     4 4.08D-03 1.16D+00 -1.35D-01 6.74D-02
#>  5     5 1.06D-03 2.79D-01 -1.69D-01 3.92D-02
#>  6     6 7.16D-04 4.96D-02 -1.14D-02 2.88D-02
#>  7     7 4.94D-04 5.01D-02 -5.27D-03 2.53D-02
#>  8     8 3.52D-04 7.38D-02 -3.24D-03 2.02D-02
#>  9     9 2.81D-04 7.26D-02 -1.86D-03 1.55D-02
#> 10    10 2.31D-04 4.46D-02 -1.03D-03 1.13D-02
#> # ℹ 39 more rows
#> 
#> [[1]][[2]]
#> # A tibble: 80 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 5.71D-04 4.74D-02 NA        4.85D-03
#>  2     2 2.53D-04 1.83D-02 -2.15D-03 2.26D-03
#>  3     3 1.15D-04 7.81D-03 -3.70D-04 1.32D-03
#>  4     4 5.76D-05 4.24D-03 -9.64D-05 9.16D-04
#>  5     5 3.65D-05 3.90D-03 -3.49D-05 6.74D-04
#>  6     6 2.51D-05 2.71D-03 -1.16D-05 5.12D-04
#>  7     7 1.88D-05 2.27D-03 -5.61D-06 3.80D-04
#>  8     8 1.55D-05 1.88D-03 -2.51D-06 3.58D-04
#>  9     9 1.14D-05 1.37D-03 -1.41D-06 3.30D-04
#> 10    10 8.52D-06 9.80D-04 -8.28D-07 3.18D-04
#> # ℹ 70 more rows
#> 
#> [[1]][[3]]
#> # A tibble: 55 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 5.06D-04 4.67D-02 NA        4.80D-03
#>  2     2 2.28D-04 2.16D-02 -2.06D-03 2.26D-03
#>  3     3 1.08D-04 8.90D-03 -3.35D-04 1.26D-03
#>  4     4 5.78D-05 5.75D-03 -8.59D-05 8.45D-04
#>  5     5 3.66D-05 3.80D-03 -3.23D-05 6.37D-04
#>  6     6 2.62D-05 3.01D-03 -1.04D-05 4.83D-04
#>  7     7 1.96D-05 2.19D-03 -4.83D-06 3.68D-04
#>  8     8 1.58D-05 1.70D-03 -2.08D-06 2.87D-04
#>  9     9 1.16D-05 1.29D-03 -1.08D-06 2.60D-04
#> 10    10 8.55D-06 9.09D-04 -5.88D-07 2.47D-04
#> # ℹ 45 more rows
#> 
#> [[1]][[4]]
#> # A tibble: 34 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.88D-04 5.34D-02 NA        2.63D-03
#>  2     2 1.29D-04 1.27D-02 -6.63D-04 1.24D-03
#>  3     3 6.55D-05 8.34D-03 -1.06D-04 7.13D-04
#>  4     4 3.69D-05 6.53D-03 -2.55D-05 4.92D-04
#>  5     5 2.02D-05 1.85D-03 -1.13D-05 3.69D-04
#>  6     6 1.43D-05 1.14D-03 -3.50D-06 2.81D-04
#>  7     7 1.07D-05 1.07D-03 -1.34D-06 2.09D-04
#>  8     8 8.58D-06 9.33D-04 -5.62D-07 1.61D-04
#>  9     9 6.34D-06 7.02D-04 -2.68D-07 1.20D-04
#> 10    10 4.70D-06 4.98D-04 -1.37D-07 9.50D-05
#> # ℹ 24 more rows
#> 
#> [[1]][[5]]
#> # A tibble: 34 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.59D-04 2.36D-02 NA        2.64D-03
#>  2     2 1.18D-04 7.84D-03 -5.92D-04 1.21D-03
#>  3     3 5.64D-05 4.13D-03 -1.05D-04 7.71D-04
#>  4     4 2.98D-05 3.92D-03 -2.67D-05 5.38D-04
#>  5     5 1.82D-05 1.97D-03 -1.04D-05 4.01D-04
#>  6     6 1.37D-05 1.46D-03 -3.39D-06 3.15D-04
#>  7     7 1.03D-05 1.15D-03 -1.56D-06 2.33D-04
#>  8     8 8.44D-06 8.97D-04 -6.73D-07 1.85D-04
#>  9     9 6.25D-06 6.66D-04 -3.38D-07 1.39D-04
#> 10    10 4.61D-06 4.73D-04 -1.76D-07 1.06D-04
#> # ℹ 24 more rows
#> 
#> [[1]][[6]]
#> # A tibble: 33 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.54D-04 2.29D-02 NA        2.60D-03
#>  2     2 1.20D-04 1.18D-02 -6.01D-04 1.24D-03
#>  3     3 5.81D-05 5.41D-03 -1.01D-04 7.15D-04
#>  4     4 3.12D-05 4.45D-03 -2.56D-05 4.87D-04
#>  5     5 1.88D-05 1.70D-03 -9.65D-06 3.71D-04
#>  6     6 1.39D-05 1.29D-03 -3.08D-06 2.79D-04
#>  7     7 1.06D-05 1.09D-03 -1.35D-06 2.12D-04
#>  8     8 8.61D-06 8.89D-04 -5.65D-07 1.67D-04
#>  9     9 6.35D-06 6.55D-04 -2.72D-07 1.27D-04
#> 10    10 4.67D-06 4.67D-04 -1.36D-07 9.69D-05
#> # ℹ 23 more rows
#> 
#> [[1]][[7]]
#> # A tibble: 37 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 3.58D-04 2.28D-02 NA        3.35D-03
#>  2     2 1.67D-04 1.48D-02 -9.82D-04 1.55D-03
#>  3     3 7.82D-05 7.75D-03 -1.67D-04 8.91D-04
#>  4     4 3.96D-05 3.92D-03 -4.34D-05 6.00D-04
#>  5     5 2.45D-05 2.21D-03 -1.55D-05 4.66D-04
#>  6     6 1.60D-05 1.60D-03 -5.26D-06 3.34D-04
#>  7     7 1.27D-05 1.19D-03 -2.18D-06 2.57D-04
#>  8     8 9.89D-06 9.69D-04 -1.00D-06 2.23D-04
#>  9     9 7.42D-06 7.12D-04 -5.42D-07 2.07D-04
#> 10    10 5.60D-06 5.12D-04 -3.15D-07 2.08D-04
#> # ℹ 27 more rows
#> 
#> [[1]][[8]]
#> # A tibble: 33 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 1.96D-04 2.33D-02 NA        1.98D-03
#>  2     2 8.84D-05 7.62D-03 -3.86D-04 9.31D-04
#>  3     3 4.50D-05 5.86D-03 -6.37D-05 5.18D-04
#>  4     4 2.63D-05 4.73D-03 -1.54D-05 3.68D-04
#>  5     5 1.50D-05 1.36D-03 -6.69D-06 2.74D-04
#>  6     6 1.09D-05 1.09D-03 -2.11D-06 2.09D-04
#>  7     7 8.21D-06 8.87D-04 -8.58D-07 1.57D-04
#>  8     8 6.64D-06 7.19D-04 -3.67D-07 1.20D-04
#>  9     9 4.98D-06 5.39D-04 -1.76D-07 9.02D-05
#> 10    10 3.69D-06 3.79D-04 -8.98D-08 7.03D-05
#> # ℹ 23 more rows
#> 
#> [[1]][[9]]
#> # A tibble: 34 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 1.95D-04 4.48D-02 NA        1.59D-03
#>  2     2 8.49D-05 7.03D-03 -2.78D-04 7.81D-04
#>  3     3 4.44D-05 6.64D-03 -4.80D-05 4.51D-04
#>  4     4 2.44D-05 4.80D-03 -1.12D-05 2.94D-04
#>  5     5 1.40D-05 1.16D-03 -5.20D-06 2.13D-04
#>  6     6 9.59D-06 6.14D-04 -1.59D-06 1.71D-04
#>  7     7 7.31D-06 4.83D-04 -5.83D-07 1.27D-04
#>  8     8 6.09D-06 4.12D-04 -2.45D-07 1.07D-04
#>  9     9 4.48D-06 3.08D-04 -1.18D-07 9.37D-05
#> 10    10 3.41D-06 2.38D-04 -6.31D-08 8.38D-05
#> # ℹ 24 more rows
#> 
#> [[1]][[10]]
#> # A tibble: 36 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 3.19D-04 1.99D-02 NA        3.38D-03
#>  2     2 1.48D-04 1.25D-02 -9.71D-04 1.56D-03
#>  3     3 7.16D-05 5.81D-03 -1.67D-04 8.26D-04
#>  4     4 3.66D-05 3.03D-03 -4.23D-05 5.88D-04
#>  5     5 2.39D-05 2.00D-03 -1.63D-05 4.52D-04
#>  6     6 1.67D-05 1.46D-03 -5.18D-06 3.29D-04
#>  7     7 1.29D-05 1.12D-03 -2.13D-06 2.40D-04
#>  8     8 1.04D-05 8.88D-04 -9.16D-07 1.84D-04
#>  9     9 7.67D-06 7.02D-04 -4.59D-07 1.41D-04
#> 10    10 5.75D-06 5.09D-04 -2.40D-07 1.25D-04
#> # ℹ 26 more rows
#> 
#> [[1]][[11]]
#> # A tibble: 33 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.11D-04 3.55D-02 NA        1.98D-03
#>  2     2 9.34D-05 8.76D-03 -3.58D-04 9.29D-04
#>  3     3 4.58D-05 6.35D-03 -5.71D-05 5.32D-04
#>  4     4 2.69D-05 4.97D-03 -1.37D-05 3.54D-04
#>  5     5 1.49D-05 1.47D-03 -6.26D-06 2.61D-04
#>  6     6 1.11D-05 1.32D-03 -2.00D-06 2.04D-04
#>  7     7 8.35D-06 9.55D-04 -7.71D-07 1.54D-04
#>  8     8 6.73D-06 7.51D-04 -3.33D-07 1.20D-04
#>  9     9 4.92D-06 5.66D-04 -1.62D-07 9.06D-05
#> 10    10 3.64D-06 4.01D-04 -8.35D-08 7.67D-05
#> # ℹ 23 more rows
#> 
#> [[1]][[12]]
#> # A tibble: 31 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.21D-04 1.62D-02 NA        2.20D-03
#>  2     2 1.04D-04 9.66D-03 -4.16D-04 1.04D-03
#>  3     3 5.17D-05 4.51D-03 -6.25D-05 5.56D-04
#>  4     4 2.83D-05 3.08D-03 -1.56D-05 3.92D-04
#>  5     5 1.75D-05 1.56D-03 -5.88D-06 2.81D-04
#>  6     6 1.23D-05 1.30D-03 -1.82D-06 2.11D-04
#>  7     7 9.02D-06 1.02D-03 -7.61D-07 1.57D-04
#>  8     8 6.97D-06 8.19D-04 -3.06D-07 1.16D-04
#>  9     9 5.02D-06 6.33D-04 -1.38D-07 8.65D-05
#> 10    10 3.64D-06 4.46D-04 -6.45D-08 6.36D-05
#> # ℹ 21 more rows
#> 
#> [[1]][[13]]
#> # A tibble: 32 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.27D-04 1.61D-02 NA        2.17D-03
#>  2     2 1.05D-04 8.94D-03 -4.15D-04 1.02D-03
#>  3     3 5.09D-05 4.37D-03 -6.41D-05 5.39D-04
#>  4     4 2.70D-05 2.79D-03 -1.63D-05 3.60D-04
#>  5     5 1.70D-05 1.61D-03 -5.94D-06 2.68D-04
#>  6     6 1.20D-05 1.36D-03 -1.84D-06 1.96D-04
#>  7     7 8.87D-06 1.01D-03 -8.10D-07 1.48D-04
#>  8     8 6.86D-06 7.97D-04 -3.26D-07 1.12D-04
#>  9     9 4.98D-06 6.16D-04 -1.50D-07 8.41D-05
#> 10    10 3.61D-06 4.33D-04 -7.19D-08 6.24D-05
#> # ℹ 22 more rows
#> 
#> [[1]][[14]]
#> # A tibble: 31 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.20D-04 1.46D-02 NA        2.24D-03
#>  2     2 1.02D-04 8.88D-03 -4.36D-04 1.05D-03
#>  3     3 4.96D-05 4.32D-03 -6.69D-05 5.53D-04
#>  4     4 2.72D-05 2.12D-03 -1.70D-05 3.60D-04
#>  5     5 1.76D-05 1.60D-03 -6.27D-06 2.67D-04
#>  6     6 1.25D-05 1.27D-03 -1.89D-06 1.92D-04
#>  7     7 9.07D-06 9.95D-04 -8.26D-07 1.44D-04
#>  8     8 6.98D-06 8.16D-04 -3.27D-07 1.09D-04
#>  9     9 5.04D-06 6.36D-04 -1.49D-07 8.03D-05
#> 10    10 3.68D-06 4.51D-04 -6.91D-08 5.98D-05
#> # ℹ 21 more rows
#> 
#> [[1]][[15]]
#> # A tibble: 31 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.25D-04 1.52D-02 NA        2.22D-03
#>  2     2 1.05D-04 9.60D-03 -4.26D-04 1.04D-03
#>  3     3 5.03D-05 4.55D-03 -6.53D-05 5.54D-04
#>  4     4 2.75D-05 2.37D-03 -1.65D-05 3.81D-04
#>  5     5 1.75D-05 1.70D-03 -6.14D-06 2.72D-04
#>  6     6 1.24D-05 1.32D-03 -1.87D-06 2.05D-04
#>  7     7 9.05D-06 1.00D-03 -8.17D-07 1.53D-04
#>  8     8 7.07D-06 8.09D-04 -3.25D-07 1.13D-04
#>  9     9 5.06D-06 6.28D-04 -1.48D-07 8.52D-05
#> 10    10 3.67D-06 4.47D-04 -6.98D-08 6.30D-05
#> # ℹ 21 more rows
#> 
#> [[1]][[16]]
#> # A tibble: 33 × 5
#>    cycle RMSDP    MaxDP    DE        OVMax   
#>    <int> <chr>    <chr>    <chr>     <chr>   
#>  1     1 2.02D-04 2.03D-02 NA        2.26D-03
#>  2     2 9.11D-05 9.97D-03 -3.93D-04 9.69D-04
#>  3     3 4.56D-05 4.84D-03 -6.94D-05 5.65D-04
#>  4     4 2.37D-05 2.35D-03 -1.80D-05 3.66D-04
#>  5     5 1.50D-05 1.35D-03 -6.17D-06 2.89D-04
#>  6     6 1.03D-05 9.93D-04 -2.01D-06 2.07D-04
#>  7     7 7.92D-06 5.97D-04 -8.57D-07 1.60D-04
#>  8     8 6.59D-06 5.36D-04 -3.55D-07 1.27D-04
#>  9     9 4.76D-06 3.48D-04 -1.74D-07 9.57D-05
#> 10    10 3.59D-06 2.66D-04 -8.83D-08 7.98D-05
#> # ℹ 23 more rows
#> 
#> 
#> [[2]]
#> [1] 1e-08
#>